First round of renaming.
