+++ Solutions/Corvus.Testing.AzureFunctions.DemoFunction/SampleFunction.cs
