+++ Solutions/Corvus.Testing.AzureFunctions.DemoFunction/host.json
