+++ Solutions/Corvus.Testing.AzureFunctions.DemoFunction/local.settings.json
