+++ Solutions/Corvus.Testing.sln
@@ -19,15 +19,17 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution 
 		.editorconfig = .editorconfig
 	EndProjectSection
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Corvus.SpecFlow.Extensions.DemoFunction", "Corvus.SpecFlow.Extensions.DemoFunction\Corvus.SpecFlow.Extensions.DemoFunction.csproj", "{144715F8-4260-40AE-9FCE-78DAA5238AC3}"
-EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Demo", "Demo", "{C07A8D88-69A4-4ABC-A057-B09D55C6B61F}"
-EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Corvus.SpecFlow.Extensions.Demo", "Corvus.SpecFlow.Extensions.Demo\Corvus.SpecFlow.Extensions.Demo.csproj", "{D70C5802-E069-4F7E-9342-DEEDD4F2502F}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Corvus.Testing.AzureFunctions", "Corvus.Testing.AzureFunctions\Corvus.Testing.AzureFunctions.csproj", "{3F86ECD4-674C-4026-9DC6-342FB43A84D6}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Corvus.Testing.AzureFunctions.Xunit.Demo", "Corvus.Testing.AzureFunctions.Xunit.Demo\Corvus.Testing.AzureFunctions.Xunit.Demo.csproj", "{BA7AE471-5DFA-4204-BB59-B4EFAC624086}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Corvus.Testing.AzureFunctions.Xunit.Demo", "Corvus.Testing.AzureFunctions.Xunit.Demo\Corvus.Testing.AzureFunctions.Xunit.Demo.csproj", "{BA7AE471-5DFA-4204-BB59-B4EFAC624086}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Corvus.Testing.AzureFunctions.SpecFlow.Demo", "Corvus.Testing.AzureFunctions.SpecFlow.Demo\Corvus.Testing.AzureFunctions.SpecFlow.Demo.csproj", "{4F726825-01C5-4BCF-B82B-E26C14565EEF}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Corvus.Testing.AzureFunctions.SpecFlow", "Corvus.Testing.AzureFunctions.SpecFlow\Corvus.Testing.AzureFunctions.SpecFlow.csproj", "{A3C7BB3D-5C1B-4A59-9D1F-CC13BAEFCF45}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Corvus.Testing.AzureFunctions.DemoFunction", "Corvus.Testing.AzureFunctions.DemoFunction\Corvus.Testing.AzureFunctions.DemoFunction.csproj", "{A82AEA52-81AA-4AAA-BEFE-8E28013E2A7A}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -43,14 +45,6 @@ Global
 		{0E316623-3DB4-4397-BFE2-43D0E4007845}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{0E316623-3DB4-4397-BFE2-43D0E4007845}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{0E316623-3DB4-4397-BFE2-43D0E4007845}.Release|Any CPU.Build.0 = Release|Any CPU
-		{144715F8-4260-40AE-9FCE-78DAA5238AC3}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{144715F8-4260-40AE-9FCE-78DAA5238AC3}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{144715F8-4260-40AE-9FCE-78DAA5238AC3}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{144715F8-4260-40AE-9FCE-78DAA5238AC3}.Release|Any CPU.Build.0 = Release|Any CPU
-		{D70C5802-E069-4F7E-9342-DEEDD4F2502F}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{D70C5802-E069-4F7E-9342-DEEDD4F2502F}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{D70C5802-E069-4F7E-9342-DEEDD4F2502F}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{D70C5802-E069-4F7E-9342-DEEDD4F2502F}.Release|Any CPU.Build.0 = Release|Any CPU
 		{3F86ECD4-674C-4026-9DC6-342FB43A84D6}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{3F86ECD4-674C-4026-9DC6-342FB43A84D6}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{3F86ECD4-674C-4026-9DC6-342FB43A84D6}.Release|Any CPU.ActiveCfg = Release|Any CPU
@@ -59,14 +53,26 @@ Global
 		{BA7AE471-5DFA-4204-BB59-B4EFAC624086}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{BA7AE471-5DFA-4204-BB59-B4EFAC624086}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{BA7AE471-5DFA-4204-BB59-B4EFAC624086}.Release|Any CPU.Build.0 = Release|Any CPU
+		{4F726825-01C5-4BCF-B82B-E26C14565EEF}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{4F726825-01C5-4BCF-B82B-E26C14565EEF}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{4F726825-01C5-4BCF-B82B-E26C14565EEF}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{4F726825-01C5-4BCF-B82B-E26C14565EEF}.Release|Any CPU.Build.0 = Release|Any CPU
+		{A3C7BB3D-5C1B-4A59-9D1F-CC13BAEFCF45}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{A3C7BB3D-5C1B-4A59-9D1F-CC13BAEFCF45}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{A3C7BB3D-5C1B-4A59-9D1F-CC13BAEFCF45}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{A3C7BB3D-5C1B-4A59-9D1F-CC13BAEFCF45}.Release|Any CPU.Build.0 = Release|Any CPU
+		{A82AEA52-81AA-4AAA-BEFE-8E28013E2A7A}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{A82AEA52-81AA-4AAA-BEFE-8E28013E2A7A}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{A82AEA52-81AA-4AAA-BEFE-8E28013E2A7A}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{A82AEA52-81AA-4AAA-BEFE-8E28013E2A7A}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
 	EndGlobalSection
 	GlobalSection(NestedProjects) = preSolution
-		{144715F8-4260-40AE-9FCE-78DAA5238AC3} = {C07A8D88-69A4-4ABC-A057-B09D55C6B61F}
-		{D70C5802-E069-4F7E-9342-DEEDD4F2502F} = {C07A8D88-69A4-4ABC-A057-B09D55C6B61F}
 		{BA7AE471-5DFA-4204-BB59-B4EFAC624086} = {C07A8D88-69A4-4ABC-A057-B09D55C6B61F}
+		{4F726825-01C5-4BCF-B82B-E26C14565EEF} = {C07A8D88-69A4-4ABC-A057-B09D55C6B61F}
+		{A82AEA52-81AA-4AAA-BEFE-8E28013E2A7A} = {C07A8D88-69A4-4ABC-A057-B09D55C6B61F}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {00493C9B-EC8E-4A4A-B07B-26D6580B6CF6}
